refactor asset page to use client component
